update pptxConverter Test
